pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "snn-encoder"
version = "0.1.0"
description = "Modular image encoding library for Spiking Neural Networks"
readme = "README.md"
requires-python = ">=3.9"
license = {text = "MIT"}

dependencies = [
    "numpy>=1.21",
    "Pillow>=9.0",
]

[project.optional-dependencies]
viz = ["matplotlib>=3.5"]
dev = ["pytest>=7.0", "matplotlib>=3.5"]

[tool.setuptools.packages.find]
include = ["snn_encoder*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
